Implemented the Booking.Application project.
Added in generic Behaviours for the Mediatr Pipeline for Logging and Validation.

Added commands for the Create / Update / Delete / Approve user stories

Added GenericTypeExtension to allow for clearer data in the LoggingBehaviour.

Added in the IBookingQueries contract and the ViewModels required to display data.

Added an IdentityService that pull the User IdentityGuid from the context.

Implemented Validators for each of the commands matching the tech spec