Implemented the Booking.Infrastructure project
Added EntityConfigurations for each of the Aggregates specifying Foreign key links.

implmented a MediatorExtensions to allow for dispating of DomainsEvents on saving entities

implemented the BookingContext with the required DbSets and functions for transactions

added in a CsvBookingContextSeed but haven't tested it yet, just in as a potential TODO.

added a db_initialmigrations.bat that when run in the cli from inside the folder builds the project and creates the migrations.